evm: rename bls files and remove g1mul, g2mul (eips PR 8945)
