+++ packages/evm/src/precompiles/10-bls12-map-fp-to-g1.ts
@@ -11,7 +11,7 @@ import { getPrecompileName } from './index.js'
 import type { EVMBLSInterface, ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
-export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
+export async function precompile10(opts: PrecompileInput): Promise<ExecResult> {
   const pName = getPrecompileName('12')
   const bls = (<any>opts._EVM)._bls! as EVMBLSInterface
