+++ packages/evm/src/precompiles/11-bls12-map-fp2-to-g2.ts
@@ -11,17 +11,17 @@ import { getPrecompileName } from './index.js'
 import type { EVMBLSInterface, ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
-export async function precompile0f(opts: PrecompileInput): Promise<ExecResult> {
-  const pName = getPrecompileName('0f')
+export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
+  const pName = getPrecompileName('13')
   const bls = (<any>opts._EVM)._bls! as EVMBLSInterface
 
   // note: the gas used is constant; even if the input is incorrect.
-  const gasUsed = opts.common.paramByEIP('bls12381G2MulGas', 2537) ?? BigInt(0)
+  const gasUsed = opts.common.paramByEIP('bls12381MapG2Gas', 2537) ?? BigInt(0)
   if (!gasLimitCheck(opts, gasUsed, pName)) {
     return OOGResult(opts.gasLimit)
   }
 
-  if (!equalityLengthCheck(opts, 288, pName)) {
+  if (!equalityLengthCheck(opts, 128, pName)) {
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INVALID_INPUT_LENGTH), opts.gasLimit)
   }
 
@@ -29,18 +29,14 @@ export async function precompile0f(opts: PrecompileInput): Promise<ExecResult> {
   const zeroByteRanges = [
     [0, 16],
     [64, 80],
-    [128, 144],
-    [192, 208],
   ]
   if (!leading16ZeroBytesCheck(opts, zeroByteRanges, pName)) {
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_POINT_NOT_ON_CURVE), opts.gasLimit)
   }
 
-  // TODO: verify that point is on G2
-
   let returnValue
   try {
-    returnValue = bls.mulG2(opts.data)
+    returnValue = bls.mapFP2toG2(opts.data)
   } catch (e: any) {
     if (opts._debug !== undefined) {
       opts._debug(`${pName} failed: ${e.message}`)
